Created new networking crate (#31)
This is incomplete and a work in progress. The goal is to create a new,
separate networking crate which has client and server plugins. Each
plugin is responsible for syncing a data model, which is public and can
be queried for in plugins outside of the Client/Server plugin.

There should be one parent entity which represents the "namespace" for
the synchronized entities (so that inspection is easy).

This crate is not finished yet, but I'm submitting this code so that
@MalekiRe can merge it and port his old client/server code to it.